Split typecheck module into more files, move errors  and op typing in submodules
